requirements.txt
telethon==1.35.0
fastapi==0.104.1
uvicorn==0.24.0
solana==0.36.7
solders==0.26.0
aiofiles==23.2.1
pydantic==2.5.0
python-dotenv==1.0.0
cryptography==45.0.4
requests==2.31.0
websockets==12.0
aiosqlite==0.19.0
jinja2==3.1.2
python-multipart==0.0.6
base58==2.1.1 